+++ src/screens/PopularMovies/components/List/index.ts
@@ -1,0 +1,3 @@
+import PopularMoviesList from './List';
+
+export default PopularMoviesList;
