--- src/screens/PopularMovies/components/PopularMoviesList/index.ts
@@ -1,3 +1,0 @@
-import PopularMoviesList from './PopularMoviesList';
-
-export default PopularMoviesList;
